feat: introduce partial classes for AzureServiceBus
